requirements.txt
fastapi==0.123.0
uvicorn[standard]==0.38.0

SQLAlchemy==2.0.44
alembic==1.17.2
psycopg[binary]==3.3.0
asyncpg==0.29.0

pydantic==2.12.5
python-dotenv==1.2.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4

redis==5.0.7
requests==2.32.3
uvloop==0.19.0
pytest-asyncio==0.23.7
httpx==0.28.1